Refactoring across backends
