+++ tensorly/backend/jax_backend.py
@@ -1,5 +1,3 @@
-import warnings
-from distutils.version import LooseVersion
 
 try:
     import jax
@@ -42,17 +40,10 @@ class JaxBackend(Backend):
     def copy(tensor):
         return copy.copy(tensor)
 
-    @staticmethod
-    def shape(tensor):
-        return tensor.shape
 
     @staticmethod
     def ndim(tensor):
         return tensor.ndim
-
-    @staticmethod
-    def dot(a, b):
-        return a.dot(b)
 
     @staticmethod
     def norm(tensor, order=2, axis=None):
@@ -69,28 +60,6 @@ class JaxBackend(Backend):
         else:
             return np.sum(np.abs(tensor)**order, axis=axis)**(1 / order)
 
-    def kr(self, matrices, weights=None, mask=None):
-        n_columns = matrices[0].shape[1]
-        n_factors = len(matrices)
-
-        start = ord('a')
-        common_dim = 'z'
-        target = ''.join(chr(start + i) for i in range(n_factors))
-        source = ','.join(i + common_dim for i in target)
-        operation = source + '->' + target + common_dim
-
-        if weights is not None:
-            matrices = [m if i else m*self.reshape(weights, (1, -1)) for i, m in enumerate(matrices)]
-
-        m = mask.reshape((-1, 1)) if mask is not None else 1
-        return np.einsum(operation, *matrices).reshape((-1, n_columns))*m
-
-    @property
-    def SVD_FUNS(self):
-        return {'numpy_svd': self.partial_svd,
-                'truncated_svd': self.truncated_svd}
-    
-    
     @staticmethod
     def sort(tensor, axis, descending = False):
         if descending:
@@ -99,8 +68,8 @@ class JaxBackend(Backend):
             return np.sort(tensor, axis=axis)
 
 for name in ['int64', 'int32', 'float64', 'float32', 'reshape', 'moveaxis',
-             'where', 'transpose', 'arange', 'ones', 'zeros',
-             'zeros_like', 'eye', 'kron', 'concatenate', 'max', 'min',
+             'where', 'dot', 'transpose', 'arange', 'ones', 'zeros',
+             'zeros_like', 'eye', 'kron', 'concatenate', 'max', 'min', 'shape',
              'all', 'mean', 'sum', 'prod', 'sign', 'abs', 'sqrt', 'argmin',
              'argmax', 'stack', 'conj', 'diag', 'clip', 'einsum']:
     JaxBackend.register_method(name, getattr(np, name))
